core:services/venv2: Add kraken to venv2 updated deps

+++ core/services/kraken/harbor/models.py
@@ -11,4 +11,4 @@ class ContainerModel(BaseModel):
 class ContainerUsageModel(BaseModel):
     cpu: float
     memory: float | str
-    disk: int | str
+    disk: float | str
